Spanion - Overwrite with SPIFlashA library
This commit copy the SPIFlashA library over SPIFlash on top of last
commit. The idea is simulate the development done in separate (not
forked repo). After this, a rebase should be possible and (hopefully)
less painful.

AMEND: rebase was painful, so, instead of the steps done in this branch,
another simples approach would be fetch SPIFlashA over this repo and
rebase it´s master on top of this master. Already done hard way now.

Signed-off-by: Josenivaldo Benito Jr <jrbenito@benito.qsl.br>